RF: Make genesis arbitraries into separate file.
