+++ fendermint/vm/genesis/src/lib.rs
@@ -14,6 +14,9 @@ use num_traits::Num;
 use serde::de::Error;
 use serde::{de, Deserialize, Serialize, Serializer};
 
+#[cfg(feature = "arb")]
+mod arb;
+
 /// Wrapper around [`Address`] to provide human readable serialization in JSON format.
 ///
 /// An alternative would be the `serde_with` crate.
@@ -168,80 +171,6 @@ pub struct Genesis {
     pub base_fee: TokenAmount,
     pub validators: Vec<Validator>,
     pub accounts: Vec<Actor>,
-}
-
-#[cfg(feature = "arb")]
-mod arb {
-    use crate::{
-        Account, Actor, ActorAddr, ActorMeta, Genesis, Multisig, Power, Validator, ValidatorKey,
-    };
-    use fendermint_testing::arb::{ArbAddress, ArbTokenAmount};
-    use fvm_shared::version::NetworkVersion;
-    use quickcheck::{Arbitrary, Gen};
-    use rand::{rngs::StdRng, SeedableRng};
-
-    impl Arbitrary for ActorMeta {
-        fn arbitrary(g: &mut Gen) -> Self {
-            if bool::arbitrary(g) {
-                ActorMeta::Account(Account {
-                    owner: ActorAddr(ArbAddress::arbitrary(g).0),
-                })
-            } else {
-                let n = u64::arbitrary(g) % 4 + 2;
-                let signers = (0..n)
-                    .map(|_| ActorAddr(ArbAddress::arbitrary(g).0))
-                    .collect();
-                let threshold = u64::arbitrary(g) % n + 1;
-                ActorMeta::MultiSig(Multisig {
-                    signers,
-                    threshold,
-                    vesting_duration: u64::arbitrary(g),
-                    vesting_start: u64::arbitrary(g),
-                })
-            }
-        }
-    }
-
-    impl Arbitrary for Actor {
-        fn arbitrary(g: &mut Gen) -> Self {
-            Self {
-                meta: ActorMeta::arbitrary(g),
-                balance: ArbTokenAmount::arbitrary(g).0,
-            }
-        }
-    }
-
-    impl Arbitrary for ValidatorKey {
-        fn arbitrary(g: &mut Gen) -> Self {
-            let mut rng = StdRng::seed_from_u64(u64::arbitrary(g));
-            let sk = libsecp256k1::SecretKey::random(&mut rng);
-            let pk = libsecp256k1::PublicKey::from_secret_key(&sk);
-            Self::new(pk)
-        }
-    }
-
-    impl Arbitrary for Validator {
-        fn arbitrary(g: &mut Gen) -> Self {
-            Self {
-                public_key: ValidatorKey::arbitrary(g),
-                power: Power(u64::arbitrary(g)),
-            }
-        }
-    }
-
-    impl Arbitrary for Genesis {
-        fn arbitrary(g: &mut Gen) -> Self {
-            let nv = usize::arbitrary(g) % 10 + 1;
-            let na = usize::arbitrary(g) % 10;
-            Self {
-                network_name: String::arbitrary(g),
-                network_version: NetworkVersion::new(*g.choose(&[18u32]).unwrap()),
-                base_fee: ArbTokenAmount::arbitrary(g).0,
-                validators: (0..nv).map(|_| Arbitrary::arbitrary(g)).collect(),
-                accounts: (0..na).map(|_| Arbitrary::arbitrary(g)).collect(),
-            }
-        }
-    }
 }
 
 #[cfg(test)]
